@@ -1,53 +1,58 @@
 {
-    "name": "secaconstrucciones",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {
-      "ng": "ng",
-      "start": "ng serve",
-      "build": "ng build --prod",
-      "test": "ng test",
-      "lint": "ng lint",
-      "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-      "@angular/animations": "^5.2.0",
-      "@angular/common": "^5.2.0",
-      "@angular/compiler": "^5.2.0",
-      "@angular/core": "^5.2.0",
-      "@angular/forms": "^5.2.0",
-      "@angular/http": "^5.2.0",
-      "@angular/platform-browser": "^5.2.0",
-      "@angular/platform-browser-dynamic": "^5.2.0",
-      "@angular/router": "^5.2.0",
-      "@ng-bootstrap/ng-bootstrap": "^2.0.0",
-      "bootstrap": "^4.1.1",
-      "core-js": "^2.4.1",
-      "jquery": "^3.3.1",
-      "ngx-bootstrap": "^2.0.5",
-      "ngx-dropdown": "0.0.22",
-      "rxjs": "^5.5.6",
-      "zone.js": "^0.8.19"
-    },
-    "devDependencies": {
-      "@angular/cli": "1.6.7",
-      "@angular/compiler-cli": "^5.2.0",
-      "@angular/language-service": "^5.2.0",
-      "@types/jasmine": "~2.8.3",
-      "@types/jasminewd2": "~2.0.2",
-      "@types/node": "~6.0.60",
-      "codelyzer": "^4.0.1",
-      "jasmine-core": "~2.8.0",
-      "jasmine-spec-reporter": "~4.2.1",
-      "karma": "~2.0.0",
-      "karma-chrome-launcher": "~2.2.0",
-      "karma-coverage-istanbul-reporter": "^1.2.1",
-      "karma-jasmine": "~1.1.0",
-      "karma-jasmine-html-reporter": "^0.2.2",
-      "protractor": "~5.1.2",
-      "ts-node": "~4.1.0",
-      "tslint": "~5.9.1",
-      "typescript": "~2.5.3"
-    }
+  "name": "secaconstrucciones",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^6.0.6",
+    "@angular/cdk": "^6.3.1",
+    "@angular/common": "^6.0.6",
+    "@angular/compiler": "^6.0.6",
+    "@angular/core": "^6.0.6",
+    "@angular/forms": "^6.0.6",
+    "@angular/http": "^6.0.6",
+    "@angular/material": "^6.3.1",
+    "@angular/platform-browser": "^6.0.6",
+    "@angular/platform-browser-dynamic": "^6.0.6",
+    "@angular/router": "^6.0.6",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
+    "bootstrap": "^4.1.1",
+    "core-js": "^2.4.1",
+    "hammerjs": "^2.0.8",
+    "jquery": "^3.3.1",
+    "ngx-bootstrap": "^3.0.1",
+    "ngx-dropdown": "0.0.22",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "zone.js": "^0.8.19"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.6",
+    "@angular/language-service": "^6.0.6",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~10.5.0",
+    "codelyzer": "^4.0.1",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^1.1.0",
+    "protractor": "~5.3.2",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.10.0",
+    "typescript": "^2.7.2"
   }
+}
